refactor(rt): move memory helper functions
